AbstractBenchmark update for fidelity option + including XGBoost
