+++ hpobench/benchmarks/ml/histgb_benchmark.py
@@ -19,10 +19,10 @@ from sklearn.metrics import accuracy_score, make_scorer
 from sklearn.experimental import enable_hist_gradient_boosting  # noqa
 from sklearn.ensemble import HistGradientBoostingClassifier
 
-from hpobench.benchmarks.ml.ml_benchmark_template import Benchmark
+from hpobench.benchmarks.ml.ml_benchmark_template import MLBenchmark
 
 
-class HistGBBenchmark(Benchmark):
+class HistGBBenchmark(MLBenchmark):
     def __init__(
             self,
             task_id: Union[int, None] = None,
